+++ examples/example.py
@@ -8,8 +8,7 @@ load_dotenv()
 API_KEY = os.getenv("RIOT_API_KEY")
 
 # We always initialize the API like this
-# We can explicitly disable url logging, by default pyke will print the url of all api calls
-api = Pyke(API_KEY, print_url=False)
+api = Pyke(API_KEY)
 
 # Every pyke method follows the same convention as the Riot API
 # For example account/v1/accounts/by-riot-id/{gameName}/{tagLine} becomes the following