+++ examples/lol_challenges.py
@@ -8,8 +8,7 @@ from pyke import Continent, Level, Pyke, Region
 load_dotenv()
 API_KEY = os.getenv("RIOT_API_KEY")
 
-# Let's not print urls, we don't want our leaderboard to be filled with random urls
-api = Pyke(API_KEY, print_url=False)
+api = Pyke(API_KEY)
 
 # Let's grab a random challenge from the lol_challenges.config method
 challenge = random.choice(api.lol_challenges.config(Region.EUW))